pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assistant-bot"
version = "0.1.0"
description = "A Telegram bot powered by OpenAI's GPT-4o model."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT License" }
authors = [
    { name = "Maxim Moroz", email = "maxim.moroz@gmail.com" }
]
keywords = ["telegram", "openai", "gpt-4", "bot", "assistant"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-telegram-bot==21.6",
    "openai==1.51.2",
    "numpy==2.0.2",
    "tiktoken==0.8.0",
    "python-dotenv==1.0.1",
]

[project.urls]
Homepage = "https://github.com/mmua/assistant_bot"

[project.scripts]
assistant_bot = "bot.bot:main"